refactor: 수정사항 반영 (2)
