+++ src/pages/notice/index.tsx
@@ -12,7 +12,7 @@ import TableContent from '@/components/notice/table/TableContent';
 import TableHead from '@/components/notice/table/TableHead';
 import { MAJOR } from '@/constants/major';
 import { PAGE_OFFSET } from '@/constants/pageoffset';
-import { useFetchNotice } from '@/query/notice';
+import { useDeleteNoticeMutation, useFetchNotice } from '@/query/notice';
 import { useFetchMe } from '@/query/user';
 import { NoticeResponse } from '@/types/notice';
 import { PATH } from '@/utils/path';
@@ -25,6 +25,7 @@ const NoticePage = () => {
   const [currentPage, setCurrentPage] = useState(1);
   const [selectedNotice, setSelectedNotice] = useState<number | null>(null);
   const { data: notices, isLoading } = useFetchNotice(MAJOR[me?.major || 0]);
+  const { mutate: deleteNoticeMutation } = useDeleteNoticeMutation();
   const sortedNotices = notices ? notices.slice().reverse() : [];
 
   const handleNoticeClick = (noticeId: number) => {
@@ -35,6 +36,18 @@ const NoticePage = () => {
     setSelectedNotice(null);
   };
 
+  const handleSubmit = () => {
+    if (selectedNoticeData) {
+      const res = confirm('정말 삭제하시겠습니까?');
+
+      if (res) {
+        deleteNoticeMutation(selectedNoticeData.id);
+        navigate(PATH.NOTICE);
+      }
+    }
+
+    handleCloseModal();
+  };
   const selectedNoticeData = selectedNotice
     ? sortedNotices.find((notice: NoticeResponse) => notice.id === selectedNotice)
     : null;
@@ -42,7 +55,8 @@ const NoticePage = () => {
   return (
     <PageTemplate>
       <Styled.Root>
-        <Styled.Title>공지사항</Styled.Title>
+          <Styled.Title>공지사항</Styled.Title>
+          <Styled.Major>{me?.major}</Styled.Major>
         {isLoading ? (
           <Loader />
         ) : (
@@ -84,6 +98,13 @@ const NoticePage = () => {
         <Styled.ModalContent>
           {selectedNoticeData && selectedNoticeData.content}
         </Styled.ModalContent>
+        {me?.is_usermode || (
+          <Styled.ModalButtons>
+            <Button variant='outlined' onClick={() => handleSubmit()}>
+              삭제
+            </Button>
+          </Styled.ModalButtons>
+        )}
       </Modal>
     </PageTemplate>
   );
