+++ src/utils/path.ts
@@ -4,7 +4,6 @@ export const PATH = {
   LOGIN: '/login',
   NOTICE: '/notice',
   NOTICE_CREATE: '/notice/create',
-  NOTICE_MANAGE: '/notice/manage',
   APPLY: '/locker/apply',
   SHARE: '/locker/share',
   SORT: '/sort',
